refactor tio ingestor class
